+++ internal/app/bindings_office.go
@@ -41,6 +41,7 @@ func (b *OfficeB) GaeaCheckUpdate() (*UpdateInfo, error) { return b.a.GaeaCheckU
 func (b *OfficeB) GaeaCheckpoints() []CheckpointMeta { return b.a.GaeaCheckpoints() }
 func (b *OfficeB) GaeaCommands() []CommandInfo { return b.a.GaeaCommands() }
 func (b *OfficeB) GaeaContext() ContextInfo { return b.a.GaeaContext() }
+func (b *OfficeB) GaeaContextNodeDetail(seq int64) (contextview.NodeDetail, error) { return b.a.GaeaContextNodeDetail(seq) }
 func (b *OfficeB) GaeaContextView() (contextview.ContextTimeline, error) { return b.a.GaeaContextView() }
 func (b *OfficeB) GaeaConvertToPdf(rel string) (ConvertPdfResult, error) { return b.a.GaeaConvertToPdf(rel) }
 func (b *OfficeB) GaeaCrossEmbed(in CrossEmbedInput) (CrossEmbedResult, error) { return b.a.GaeaCrossEmbed(in) }